// src/shared/enums/core/error-codes.enum.ts
export enum ErrorCodes {
  UNAUTHORIZED = "UNAUTHORIZED",
  VALIDATION_ERROR = "VALIDATION_ERROR",
  FORBIDDEN = "FORBIDDEN",
  RESOURCE_ALREADY_EXISTS = "RESOURCE_ALREADY_EXISTS",
  RESOURCE_NOT_FOUND = "RESOURCE_NOT_FOUND",
  BAD_REQUEST = "BAD_REQUEST",
  LICENSE_INACTIVE = "LICENSE_INACTIVE",
  TOO_MANY_REQUESTS = "TOO_MANY_REQUESTS",
  SESSION_LIMIT_REACHED = "SESSION_LIMIT_REACHED",
  PAYMENT_GATEWAY_ERROR = "PAYMENT_GATEWAY_ERROR",
  STORAGE_PROVIDER_ERROR = "STORAGE_PROVIDER_ERROR",
  DATABASE_ERROR = "DATABASE_ERROR",
  BUSINESS_DAY_CLOSED = "BUSINESS_DAY_CLOSED",
  PREVIOUS_BUSINESS_DAY_OPEN = "PREVIOUS_BUSINESS_DAY_OPEN",
}